Run the bridge's modal sections in double precision
A direct-form section's coefficients sit within theta^2 of 2 and 1, so a
float held the bridge's lowest modes' pole angles to about 5% at 384 kHz.
That turned the heave's fundamental 93 degrees against the rocking moment
and changed a microphone's low E by up to 6 dB with the host rate; at 96
and 192 kHz it already moved 1-2 dB. In double precision every rate agrees
to 0.1 degree and 0.5% of sustain level, and all suites stay green.

+++ Source/DSP/AcustraEngine.h
@@ -346,20 +346,24 @@ private:
         void write(float value) noexcept;
     };
 
+    // Double precision: a direct-form section's coefficients sit within
+    // theta^2 of 2 and 1, and at 384 kHz a float held the bridge's air-mode
+    // group's pole angles to only about 5%, which turned the heave's H1 by
+    // 90 degrees against the rock.
     struct BridgeMode
     {
-        float denominator1 { 0.0f };
-        float denominator2 { 0.0f };
-        float numerator1 { 0.0f };
-        float numerator2 { 0.0f };
-        float input1 { 0.0f };
-        float output1 { 0.0f };
-        float output2 { 0.0f };
+        double denominator1 { 0.0 };
+        double denominator2 { 0.0 };
+        double numerator1 { 0.0 };
+        double numerator2 { 0.0 };
+        double input1 { 0.0 };
+        double output1 { 0.0 };
+        double output2 { 0.0 };
 
-        float processPast(float input) noexcept;
+        double processPast(double input) noexcept;
         void reset() noexcept
         {
-            input1 = output1 = output2 = 0.0f;
+            input1 = output1 = output2 = 0.0;
         }
     };
